+++ api/python/quilt3/_graphql_client/fragments.py
@@ -7,7 +7,7 @@ from typing import Annotated, Any, List, Literal, Optional, Union
 from pydantic import Field
 
 from .base_model import BaseModel
-from .enums import APIKeyStatus
+from .enums import APIKeyStatus, BucketPermissionLevel
 
 
 class APIKeySelection(BaseModel):
@@ -49,10 +49,51 @@ class InvalidInputSelectionErrors(BaseModel):
     context: Optional[Any]
 
 
+class PermissionSelection(BaseModel):
+    bucket: "PermissionSelectionBucket"
+    level: BucketPermissionLevel
+
+
+class PermissionSelectionBucket(BaseModel):
+    name: str
+
+
+class PolicySummarySelection(BaseModel):
+    id: str
+    title: str
+    arn: str
+    managed: bool
+    permissions: List["PolicySummarySelectionPermissions"]
+
+
+class PolicySummarySelectionPermissions(PermissionSelection):
+    pass
+
+
+class RoleBucketPermissionSelection(BaseModel):
+    bucket: "RoleBucketPermissionSelectionBucket"
+    level: BucketPermissionLevel
+
+
+class RoleBucketPermissionSelectionBucket(BaseModel):
+    name: str
+
+
 class ManagedRoleSelection(BaseModel):
+    typename__: str = Field(alias="__typename")
     id: str
     name: str
     arn: str
+    policies: List["ManagedRoleSelectionPolicies"]
+    permissions: List["ManagedRoleSelectionPermissions"]
+
+
+class ManagedRoleSelectionPolicies(PolicySummarySelection):
+    pass
+
+
+class ManagedRoleSelectionPermissions(RoleBucketPermissionSelection):
+    pass
 
 
 class OperationErrorSelection(BaseModel):
@@ -61,7 +102,25 @@ class OperationErrorSelection(BaseModel):
     context: Optional[Any]
 
 
+class PolicySelection(BaseModel):
+    id: str
+    title: str
+    arn: str
+    managed: bool
+    permissions: List["PolicySelectionPermissions"]
+    roles: List["PolicySelectionRoles"]
+
+
+class PolicySelectionPermissions(PermissionSelection):
+    pass
+
+
+class PolicySelectionRoles(ManagedRoleSelection):
+    pass
+
+
 class UnmanagedRoleSelection(BaseModel):
+    typename__: str = Field(alias="__typename")
     id: str
     name: str
     arn: str
@@ -122,8 +181,12 @@ class SsoConfigSelectionUploader(UserSelection):
 APIKeySelection.model_rebuild()
 BucketConfigSelection.model_rebuild()
 InvalidInputSelection.model_rebuild()
+PermissionSelection.model_rebuild()
+PolicySummarySelection.model_rebuild()
+RoleBucketPermissionSelection.model_rebuild()
 ManagedRoleSelection.model_rebuild()
 OperationErrorSelection.model_rebuild()
+PolicySelection.model_rebuild()
 UnmanagedRoleSelection.model_rebuild()
 UserSelection.model_rebuild()
 SsoConfigSelection.model_rebuild()