+++ catalog/app/website/pages/Install/Install.jsx
@@ -12,7 +12,7 @@ const MP_LINK =
   'https://aws.amazon.com/marketplace/pp/Quilt-Data-Quilt-Business/B07QF1VXFQ'
 const PREREQ_LINK =
   'https://docs.quilt.bio/references/technical-reference#before-you-install-quilt'
-const MAILTO = 'mailto:contact@quiltdata.com'
+const MAILTO = 'mailto:support@quilt.bio'
 
 const useStyles = M.makeStyles((t) => ({
   container: {